burr/tracking/server/requirements-s3.txt
aerich
aiobotocore
fastapi-utils
tortoise-orm[accel, asyncmy]
